change to float32_t for operating variables, as STM32G4 FPU supports single precision floating data only

+++ mc/include/filters.h
@@ -12,18 +12,19 @@
 extern "C" {
 #endif
 
+#include "commontypes.h"
 
 // y(k) = b_1 y(k-1) + a_0 x(k) + a_1 x(k-1)
 //
 typedef struct
 {
     // filter-1st data
-    double   y;
-    double   u;
+    float32_t   y;
+    float32_t   u;
     // filter-1st param
-    double   a0;
-    double   a1;
-    double   b1;
+    float32_t   a0;
+    float32_t   a1;
+    float32_t   b1;
 } Lpf1st_Obj_t;
  
  
@@ -37,8 +38,8 @@ typedef struct
  *                  $y(k) = b_1 y(k-1) + a_0 x(k) + a_1 x(k-1)$ 
  *                  usually $a_0 + a_1 + b_1 = 1$ to have unity gain.
  */
-void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, double y, double u,
-                double a0, double a1, double b1);
+void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, float32_t y, float32_t u,
+                float32_t a0, float32_t a1, float32_t b1);
 
 /**
  * @brief      first order lowpass filter update
@@ -47,7 +48,7 @@ void lpf_1st_init(Lpf1st_Obj_t* const lpf_1st_inst, double y, double u,
  * 
  *     Implment a general first order lowpass filter
  */
-void lpf_1st_update(Lpf1st_Obj_t* const lpf_1st_inst, double u);
+void lpf_1st_update(Lpf1st_Obj_t* const lpf_1st_inst, float32_t u);
 
 #ifdef __cplusplus
 }
